requirements.txt
# OpenAI SDK
openai>=1.12.0

# Environment management
python-dotenv>=1.0.0

# Audio processing (for later phases)
# pydub>=0.25.1

# HTTP requests
requests>=2.31.0
